javamail-configuration.xsd: make readmailhost port be an xs:int type instead of xs:integer.

+++ opennms-config-model/src/main/java/org/opennms/netmgt/config/javamail/ReadmailHost.java
@@ -73,7 +73,7 @@ import org.opennms.core.xml.ValidateUsing;
      * Field _port.
      */
     @XmlAttribute(name="port")
-    private Long _port;
+    private Integer _port;
 
     /**
      * Basically any attributes that help setup the javamailer's
@@ -161,7 +161,7 @@ import org.opennms.core.xml.ValidateUsing;
      *
      * @return the value of field 'Port'.
      */
-    public Long getPort(
+    public Integer getPort(
     ) {
         return this._port == null ? 110 : this._port;
     }
@@ -281,7 +281,7 @@ import org.opennms.core.xml.ValidateUsing;
      * @param port the value of field 'port'.
      */
     public void setPort(
-            final Long port) {
+            final Integer port) {
         this._port = port;
     }
